Final changes, there are some issues with exiting keybinds and input / drawing could be improved but the logic is correct

+++ src/dft.hpp
@@ -6,8 +6,8 @@
 
 namespace dft {
 
-    using DFTResult = std::vector<std::tuple<std::complex<float>, float, float, float>>;
+    using DFTResult = std::vector<std::tuple<std::complex<double>, double, double, double>>;
 
-    DFTResult compute(const std::vector<std::complex<float>>& samples);
+    DFTResult compute(const std::vector<std::complex<double>>& samples);
 
 }  // namespace dft
